+++ BitmovinConvivaAnalytics/Classes/Helper/BitmovinPlayerListener.swift
@@ -10,10 +10,10 @@ import Foundation
 import BitmovinPlayer
 
 class BitmovinPlayerListener: NSObject {
-    let player: BitmovinPlayer
+    let player: Player
     weak var delegate: BitmovinPlayerListenerDelegate?
 
-    init(player: BitmovinPlayer) {
+    init(player: Player) {
         self.player = player
         super.init()
         self.player.add(listener: self)
@@ -25,11 +25,11 @@ class BitmovinPlayerListener: NSObject {
 }
 
 extension BitmovinPlayerListener: PlayerListener {
-    func onEvent(_ event: PlayerEvent) {
+    func onEvent(_ event: Event, player: Player) {
         delegate?.onEvent(event)
     }
 
-    func onSourceUnloaded(_ event: SourceUnloadedEvent) {
+    func onSourceUnloaded(_ event: SourceUnloadedEvent, player: Player) {
         // The default SDK error handling is that it triggers the onSourceUnloaded before the onError event.
         // To track errors to conviva we need to delay the onSourceUnloaded to ensure the onError event is
         // called first.
@@ -38,92 +38,101 @@ extension BitmovinPlayerListener: PlayerListener {
         }
     }
 
-    func onTimeChanged(_ event: TimeChangedEvent) {
+    func onTimeChanged(_ event: TimeChangedEvent, player: Player) {
         delegate?.onTimeChanged()
     }
 
-    func onError(_ event: ErrorEvent) {
-        delegate?.onError(event)
+    func onPlayerError(_ event: PlayerErrorEvent, player: Player) {
+        delegate?.onPlayerError(event)
     }
 
-    func onMuted(_ event: MutedEvent) {
+    func onSourceError(_ event: SourceErrorEvent, player: Player) {
+        delegate?.onSourceError(event)
+    }
+
+    func onMuted(_ event: MutedEvent, player: Player) {
         delegate?.onMuted(event)
     }
 
-    func onUnmuted(_ event: UnmutedEvent) {
+    func onUnmuted(_ event: UnmutedEvent, player: Player) {
         delegate?.onUnmuted(event)
     }
 
     // MARK: - Playback state events
-    func onPlay(_ event: PlayEvent) {
+    func onPlay(_ event: PlayEvent, player: Player) {
         delegate?.onPlay()
     }
 
-    func onPlaying(_ event: PlayingEvent) {
+    func onPlaying(_ event: PlayingEvent, player: Player) {
         delegate?.onPlaying()
     }
 
-    func onPaused(_ event: PausedEvent) {
+    func onPaused(_ event: PausedEvent, player: Player) {
         delegate?.onPaused()
     }
 
-    func onPlaybackFinished(_ event: PlaybackFinishedEvent) {
+    func onPlaylistTransition(_ event: PlaylistTransitionEvent, player: Player) {
+        // playlist is not supported yet, finish tracking for now
         delegate?.onPlaybackFinished()
     }
 
-    func onStallStarted(_ event: StallStartedEvent) {
+    func onPlaybackFinished(_ event: PlaybackFinishedEvent, player: Player) {
+        delegate?.onPlaybackFinished()
+    }
+
+    func onStallStarted(_ event: StallStartedEvent, player: Player) {
         delegate?.onStallStarted()
     }
 
-    func onStallEnded(_ event: StallEndedEvent) {
+    func onStallEnded(_ event: StallEndedEvent, player: Player) {
         delegate?.onStallEnded()
     }
 
     // MARK: - Seek / Timeshift events
-    func onSeek(_ event: SeekEvent) {
+    func onSeek(_ event: SeekEvent, player: Player) {
         delegate?.onSeek(event)
     }
 
-    func onSeeked(_ event: SeekedEvent) {
+    func onSeeked(_ event: SeekedEvent, player: Player) {
         delegate?.onSeeked()
     }
 
-    func onTimeShift(_ event: TimeShiftEvent) {
+    func onTimeShift(_ event: TimeShiftEvent, player: Player) {
         delegate?.onTimeShift(event)
     }
 
-    func onTimeShifted(_ event: TimeShiftedEvent) {
+    func onTimeShifted(_ event: TimeShiftedEvent, player: Player) {
         delegate?.onTimeShifted()
     }
 
     #if !os(tvOS)
     // MARK: - Ad events
-    func onAdStarted(_ event: AdStartedEvent) {
+    func onAdStarted(_ event: AdStartedEvent, player: Player) {
         delegate?.onAdStarted(event)
     }
 
-    func onAdFinished(_ event: AdFinishedEvent) {
+    func onAdFinished(_ event: AdFinishedEvent, player: Player) {
         delegate?.onAdFinished()
     }
 
-    func onAdSkipped(_ event: AdSkippedEvent) {
+    func onAdSkipped(_ event: AdSkippedEvent, player: Player) {
         delegate?.onAdSkipped(event)
     }
 
-    func onAdError(_ event: AdErrorEvent) {
+    func onAdError(_ event: AdErrorEvent, player: Player) {
         delegate?.onAdError(event)
     }
-    
-    func onAdBreakStarted(_ event : AdBreakStartedEvent) {
+
+    func onAdBreakStarted(_ event: AdBreakStartedEvent, player: Player) {
         delegate?.onAdBreakStarted(event)
     }
-    
-    func onAdBreakFinished(_ event : AdBreakFinishedEvent) {
+
+    func onAdBreakFinished(_ event: AdBreakFinishedEvent, player: Player) {
         delegate?.onAdBreakFinished(event)
     }
     #endif
 
-    func onDestroy(_ event: DestroyEvent) {
+    func onDestroy(_ event: DestroyEvent, player: Player) {
         delegate?.onDestroy()
     }
 }