+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0]
+
+### Added
+
+- This is first version using Bitmovin Player iOS V3 SDK. This is not backward compatible with V2 player SDK.
+
+### Known Issues
+- Playlist API available in V3 player SDK is not supported.
+
 ## [1.4.0]
 
 ### Fixed